pyproject.toml
[project]
name = "google-agemt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.3.0",
]
